+++ src/icons.tsx
@@ -4,8 +4,8 @@ function Icon({ path }: { path: ReactNode }) {
   return (
     <svg
       style={{ display: "block" }}
-      width="16"
-      height="16"
+      width="20"
+      height="20"
       viewBox="0 0 24 24"
       fill="none"
       stroke="currentColor"